make cheat cvars allowed in single player by default. fix an issue with referenced packs getting forgotten when running non-dedicated servers. whichpack now always references packs even if allow_download_refpackages is disabled. add setcursormode for compat with dp's csqc. fix some qcc issues that came up. bump some q3bsp sanity limits a little. greater precision for android's vibrate feature.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4567 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/sys.h
@@ -89,7 +89,7 @@ void Sys_SendKeyEvents (void);
 
 int Sys_EnumerateFiles (const char *gpath, const char *match, int (QDECL *func)(const char *fname, int fsize, void *parm, searchpathfuncs_t *spath), void *parm, searchpathfuncs_t *spath);
 
-void Sys_Vibrate(int count);
+void Sys_Vibrate(float count);
 
 qboolean Sys_GetDesktopParameters(int *width, int *height, int *bpp, int *refreshrate);
